bug fixed

+++ horgand/src/organDSPEffects.C
@@ -81,7 +81,7 @@ HOR::Effect_Chorus()
 
       ldelay=ms+ch_delay+Chorus_LFO(&Chorus_X_L);
       dell=(ldelay1*(PERIOD-i)+ldelay*i)/PERIOD;
-      elkel=cl_counter-ceilf(dell-1.0);
+      elkel=cl_counter-(int)ceilf(dell-1.0);
       if (elkel<0) elkel +=8192;
       if (elkel>=8192) elkel -=8192;
       elkel2=elkel-1;
@@ -97,7 +97,7 @@ HOR::Effect_Chorus()
       // R Channel
       rdelay=ms+ch_delay+Chorus_LFO(&Chorus_X_R);
       dell=(rdelay1*(PERIOD-i)+rdelay*i)/PERIOD;
-      elkel=cl_counter-ceilf(dell-1.0);
+      elkel=cl_counter-(int)ceilf(dell-1.0);
       if (elkel<0) elkel +=8192;
       if (elkel>=8192) elkel -=8192;
       elkel2=elkel-1;
